refactor(proxy): introduce Runtime interface for backend strategies
Prep for adding a SwiftLM/MLX backend alongside llama-server. Extracts the
llama-specific pieces of the proxy manager behind a Runtime interface so
ModelManager can drive multiple backend types through one lifecycle.

- Add Runtime interface (BinaryPath, WorkingDir, BuildArgs, HealthURL,
  IsStartupError) in internal/proxy/backend.go
- Add LlamaRuntime in internal/proxy/backend_llama.go holding the logic
  lifted from manager.go: arg building, mmproj discovery, startup-error
  log scanning, buildLlamaServerArgs helper
- Backend struct gains Kind and Runtime fields, set at creation time
- selectRuntime is a stub that always returns LlamaRuntime; a follow-up
  commit will dispatch based on metadata.yaml

No behavior change. All tests and lints pass.
